% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dt_id2poly}
\alias{dt_id2poly}
\title{dt_id2poly}
\usage{
dt_id2poly(df)
}
\arguments{
\item{df}{A data.table with the columns of \verb{"date", "x", "y", "id", "value"},
\code{value} is the anomaly.}
}
\description{
dt_id2poly
}
