% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST3D-package.R
\docType{package}
\name{ST3D-package}
\alias{ST3D}
\alias{ST3D-package}
\title{ST3D: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
